requirements.txt
Package                       Version
----------------------------- --------
aiohttp                       3.8.4
aiosignal                     1.3.1
asgiref                       3.6.0
async-timeout                 4.0.2
attrs                         23.1.0
certifi                       2023.5.7
channels                      4.0.0
channels-redis                4.1.0
charset-normalizer            3.1.0
colorama                      0.4.6
coverage                      7.2.3
Django                        4.2
django-cors-headers           3.14.0
django-filter                 23.1
django-redis                  5.2.0
django-redis-cache            3.0.1
djangorestframework           3.14.0
djangorestframework-simplejwt 5.2.2
drf-nested-routers            0.93.4
frozenlist                    1.3.3
gunicorn                      20.1.0
idna                          3.4
iniconfig                     2.0.0
msgpack                       1.0.7
multidict                     6.0.4
openai                        0.27.7
packaging                     23.1
Pillow                        9.5.0
pip                           23.1.2
pluggy                        1.0.0
psycopg2-binary               2.9.6
PyJWT                         2.6.0
pytest                        7.3.1
pytest-cov                    4.0.0
pytest-django                 4.5.2
python-dotenv                 1.0.0
pytz                          2023.3
redis                         4.5.5
requests                      2.31.0
setuptools                    67.6.0
sqlparse                      0.4.3
tqdm                          4.65.0
tzdata                        2023.3
urllib3                       2.0.2
wheel                         0.40.0
yarl                          1.9.2
